@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import Crisis from '../model/crisis';
 import { CrisisService } from './../service/crisis.service';
